[clang-tidy] Address Copilot review comments on type-template arguments
mipi_spi/ili9xxx watermark updates: x/y are int params; using
std::min<uint16_t> would cast x/y to uint16_t before comparing,
wrapping/truncating if the input ever leaves [0, 65535]. Switch to
std::min<int>/std::max<int> to match the original promotion semantics;
the narrowing to uint16_t happens implicitly on assignment as before.

esp32_touch queue_size is a size_t; std::max<uint32_t> would truncate
on 64-bit hosts. Use std::max(queue_size, size_t{8}) to stay in
the natural type.

addressable_light shift amount: this->size() is size_t; clamp in size_t
and cast the result back to int32_t once.

+++ esphome/components/ili9xxx/ili9xxx_display.cpp
@@ -204,10 +204,10 @@ void HOT ILI9XXXDisplay::draw_absolute_pixel_internal(int x, int y, Color color)
   }
   if (updated) {
     // low and high watermark may speed up drawing from buffer
-    this->x_low_ = std::min<uint16_t>(this->x_low_, x);
-    this->y_low_ = std::min<uint16_t>(this->y_low_, y);
-    this->x_high_ = std::max<uint16_t>(this->x_high_, x);
-    this->y_high_ = std::max<uint16_t>(this->y_high_, y);
+    this->x_low_ = std::min<int>(this->x_low_, x);
+    this->y_low_ = std::min<int>(this->y_low_, y);
+    this->x_high_ = std::max<int>(this->x_high_, x);
+    this->y_high_ = std::max<int>(this->y_high_, y);
   }
 }
